veeva-data-quality-system/requirements.txt
# Core database and data processing
SQLAlchemy>=2.0.0
pandas>=2.0.0
# sqlite3 is built into Python

# Data validation and quality
great-expectations>=0.17.0
pydantic>=2.0.0

# CLI and utilities  
click>=8.0.0
rich>=13.0.0
tabulate>=0.9.0

# Configuration management
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Testing framework
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Code quality and formatting
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Performance and monitoring
psutil>=5.9.0
memory-profiler>=0.60.0

# Date and time handling
python-dateutil>=2.8.0

# Logging and debugging
loguru>=0.7.0

# Development utilities
ipython>=8.0.0
jupyter>=1.0.0